# docs/admin/swagger.yaml
basePath: /api/admin
definitions:
  base_info.Success:
    properties:
      code:
        example: 200
        type: integer
      data: {}
      msg:
        example: err msg
        type: string
      reason:
        description: 错误原因，定义为业务判定错误码
        example: success
        type: string
    type: object
  base_info.Swagger400Resp:
    properties:
      code:
        example: 400
        type: integer
      msg:
        example: err msg
        type: string
      reason:
        example: err_reason
        type: string
    type: object
  base_info.Swagger401Resp:
    properties:
      code:
        example: 401
        type: integer
      msg:
        example: err msg
        type: string
      reason:
        example: err_reason
        type: string
    type: object
  base_info.Swagger500Resp:
    properties:
      code:
        example: 500
        type: integer
      msg:
        example: err msg
        type: string
      reason:
        example: err_reason
        type: string
    type: object
...
  commands.AssignTenantPermissionsCommand:
    properties:
      permissionIds:
        items:
          type: integer
        type: array
      tenantId:
        type: string
    required:
    - permissionIds
    - tenantId
    type: object
...
  commands.CreateTenantCommand:
    properties:
      adminUser:
        $ref: '#/definitions/commands.CreateUserCommand'
      code:
        type: string
      description:
        maxLength: 200
        type: string
      expireTime:
        type: integer
      isDefault:
        enum:
        - 0
        - 1
        type: integer
      name:
        type: string
    required:
    - adminUser
    - code
    - name
    type: object
  commands.CreateUserCommand:
    properties:
      avatar:
        description: 头像
        type: string
      email:
        description: 邮箱
        type: string
      name:
        description: 名称
        type: string
      nickname:
        description: 昵称
        type: string
      password:
        description: 密码
        minLength: 6
        type: string
      phone:
        description: 手机号
        type: string
      roleIds:
        description: 角色ID列表
        items:
          type: integer
        type: array
      username:
        description: 用户名
        type: string
    required:
    - password
    - username
    type: object
...
  commands.UpdateTenantCommand:
    properties:
      description:
        maxLength: 200
        type: string
      expireTime:
        type: integer
      id:
        type: string
      isDefault:
        enum:
        - 0
        - 1
        type: integer
      name:
        type: string
    required:
    - id
    type: object
...
  dto.PermissionsDto:
    properties:
      children:
        description: 子权限
        items:
          $ref: '#/definitions/dto.PermissionsDto'
        type: array
      code:
        description: 编码
        type: string
      createdAt:
        description: 创建时间
        type: integer
      description:
        description: 描述
        type: string
      icon:
        description: 图标
        type: string
      id:
        description: ID
        type: integer
      localize:
        description: 本地化
        type: string
      name:
        description: 名称
        type: string
      parentId:
        description: 父级ID
        type: integer
      parentPath:
        description: 父级路径
        type: string
      path:
        description: 路径
        type: string
      properties:
        description: 属性
        type: string
      resources:
        description: 资源列表
        items:
          $ref: '#/definitions/dto.PermissionsResourceDto'
        type: array
      sequence:
        description: 排序
        type: integer
      status:
        description: 状态
        type: integer
      type:
        description: 类型
        type: integer
      updatedAt:
        description: 更新时间
        type: integer
    type: object
  dto.PermissionsResourceDto:
    properties:
      method:
        description: HTTP方法
        type: string
      path:
        description: 资源路径
        type: string
    type: object
...
  dto.TenantDto:
    properties:
      adminUser:
        allOf:
        - $ref: '#/definitions/dto.UserDto'
        description: 管理员用户
      code:
        description: 租户编码
        type: string
      createdAt:
        description: 创建时间
        type: integer
      description:
        description: 描述
        type: string
      domain:
        description: 域名
        type: string
      expireTime:
        description: 过期时间
        type: integer
      id:
        description: ID
        type: string
      isDefault:
        description: 是否默认租户
        type: integer
      name:
        description: 租户名称
        type: string
      status:
        description: 状态
        type: integer
      updatedAt:
        description: 更新时间
        type: integer
    type: object
  dto.UserDto:
    properties:
      avatar:
        description: 头像
        type: string
      createdAt:
        description: 创建时间
        type: integer
      email:
        description: 邮箱
        type: string
      id:
        description: 用户ID
        type: string
      invitationCode:
        description: 邀请码
        type: string
      name:
        description: 姓名
        type: string
      nickname:
        description: 昵称
        type: string
      phone:
        description: 手机号
        type: string
      remark:
        description: 备注
        type: string
      roleIds:
        description: 角色ID列表
        items:
          type: integer
        type: array
      status:
        description: 状态,1启用,2禁用
        type: integer
      tenantId:
        description: 租户ID
        type: string
      updatedAt:
        description: 更新时间
        type: integer
      username:
        description: 用户名
        type: string
    type: object
...
  models.PageRes-dto_TenantDto:
    properties:
      list:
        items:
          $ref: '#/definitions/dto.TenantDto'
        type: array
      total:
        type: integer
    type: object
...
host: localhost:8888
info:
  contact:
    name: go-server-template-admin
  description: This is a demo using go-server-template-admin.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: ares-ddd-admin
  version: "1.0"
paths:
...
  /v1/sys/tenant:
    get:
      consumes:
      - application/json
      description: 获取租户列表，支持分页和条件查询
      operationId: TenantList
      parameters:
      - in: query
        name: code
        type: string
      - description: 页码，从1开始
        in: query
        name: current
        type: integer
      - in: query
        name: name
        type: string
      - description: 页码大小，最大100
        in: query
        name: size
        type: integer
      - in: query
        name: status
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/base_info.Success'
            - properties:
                data:
                  $ref: '#/definitions/models.PageRes-dto_TenantDto'
              type: object
        "400":
          description: code为400 参数输入错误
          schema:
            $ref: '#/definitions/base_info.Swagger400Resp'
        "401":
          description: code为401 token未带上
          schema:
            $ref: '#/definitions/base_info.Swagger401Resp'
        "500":
          description: code为500 服务端内部错误
          schema:
            $ref: '#/definitions/base_info.Swagger500Resp'
      summary: 获取租户列表
      tags:
      - 系统租户
    post:
      consumes:
      - application/json
      description: 添加租户，包括管理员用户信息
      operationId: AddTenant
      parameters:
      - description: 租户创建信息
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/commands.CreateTenantCommand'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/base_info.Success'
        "400":
          description: code为400 参数输入错误
          schema:
            $ref: '#/definitions/base_info.Swagger400Resp'
        "401":
          description: code为401 token未带上
          schema:
            $ref: '#/definitions/base_info.Swagger401Resp'
        "500":
          description: code为500 服务端内部错误
          schema:
            $ref: '#/definitions/base_info.Swagger500Resp'
      summary: 添加租户
      tags:
      - 系统租户
    put:
      consumes:
      - application/json
      description: 更新租户信息
      operationId: UpdateTenant
      parameters:
      - description: 租户更新信息
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/commands.UpdateTenantCommand'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/base_info.Success'
        "400":
          description: code为400 参数输入错误
          schema:
            $ref: '#/definitions/base_info.Swagger400Resp'
        "401":
          description: code为401 token未带上
          schema:
            $ref: '#/definitions/base_info.Swagger401Resp'
        "500":
          description: code为500 服务端内部错误
          schema:
            $ref: '#/definitions/base_info.Swagger500Resp'
      summary: 更新租户
      tags:
      - 系统租户
  /v1/sys/tenant/{id}:
    delete:
      consumes:
      - application/json
      description: 删除指定ID的租户
      operationId: DeleteTenant
      parameters:
      - description: 租户ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/base_info.Success'
        "400":
          description: code为400 参数输入错误
          schema:
            $ref: '#/definitions/base_info.Swagger400Resp'
        "401":
          description: code为401 token未带上
          schema:
            $ref: '#/definitions/base_info.Swagger401Resp'
        "500":
          description: code为500 服务端内部错误
          schema:
            $ref: '#/definitions/base_info.Swagger500Resp'
      summary: 删除租户
      tags:
      - 系统租户
  /v1/sys/tenant/permissions:
    put:
      consumes:
      - application/json
      description: 为指定租户分配权限
      operationId: AssignTenantPermissions
      parameters:
      - description: 权限分配信息
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/commands.AssignTenantPermissionsCommand'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/base_info.Success'
        "400":
          description: code为400 参数输入错误
          schema:
            $ref: '#/definitions/base_info.Swagger400Resp'
        "401":
          description: code为401 token未带上
          schema:
            $ref: '#/definitions/base_info.Swagger401Resp'
        "500":
          description: code为500 服务端内部错误
          schema:
            $ref: '#/definitions/base_info.Swagger500Resp'
      summary: 分配权限给租户
      tags:
      - 系统租户
  /v1/sys/tenant/permissions/:id:
    get:
      consumes:
      - application/json
      description: 获取指定租户的权限列表
      operationId: GetTenantPermissions
      parameters:
      - description: 租户ID
        in: query
        name: tenant_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/base_info.Success'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.PermissionsDto'
                  type: array
              type: object
        "400":
          description: code为400 参数输入错误
          schema:
            $ref: '#/definitions/base_info.Swagger400Resp'
        "401":
          description: code为401 token未带上
          schema:
            $ref: '#/definitions/base_info.Swagger401Resp'
        "500":
          description: code为500 服务端内部错误
          schema:
            $ref: '#/definitions/base_info.Swagger500Resp'
      summary: 获取租户权限
      tags:
      - 系统租户
...
schemes:
- http
swagger: "2.0"